pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logreader"
version = "0.1.0"
description = "FRC Robot Log Reader — reads and processes WPILib .wpilog and CTRE .hoot log files"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LogReader1 Team"},
]
readme = "README.md"

dependencies = [
    "robotpy-wpiutil",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
logreader = "logreader.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
